requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
cloudinary>=1.40.0
fastapi-limiter>=0.1.6
redis>=5.0.0
fastapi-mail>=1.4.1
jinja2>=3.1.0
